+++ one/alf/cache.py
@@ -1,8 +1,5 @@
 """Construct Parquet database from local file system.
-TODO Consolidate functions with ALF.files
 TODO Deal graciously with empty tables
-TODO Support Subjects folder as optional
-TODO Allow lab passed as arg instead of folder structure
 """
 
 
@@ -14,13 +11,17 @@ import datetime
 import uuid
 from functools import partial
 from pathlib import Path
-import re
+import warnings
 
 import pandas as pd
-
 from iblutil.io import parquet
 from iblutil.io.hashfile import md5
 
+from one.alf.io import iter_sessions
+from one.alf.files import session_path_parts, get_alf_path
+from one.alf.spec import is_valid
+
+__all__ = ['make_parquet_db']
 
 # -------------------------------------------------------------------------------------------------
 # Global variables
@@ -41,172 +42,46 @@ DATASETS_COLUMNS = (
     'eid',              # int64
     'session_path',     # relative to the root
     'rel_path',         # relative to the session path, includes the filename
-    'file_size',
+    'file_size',        # file size in bytes
     'hash',             # sha1/md5, computed in load function
     'exists',           # bool
 )
-
-
-def _compile(r):
-    r = r.replace('/', r'\/')
-    return re.compile(r)
-
-
-def _pattern_to_regex(pattern):
-    """Convert a path pattern with {...} into a regex."""
-    return _compile(re.sub(r'{(\w+)}', r'(?P<\1>[a-zA-Z0-9\_\-\.]+)', pattern))
-
-
-SESSION_PATTERN = "{lab}/Subjects/{subject}/{date}/{number}"
-SESSION_REGEX = _pattern_to_regex('^%s/?$' % SESSION_PATTERN)
-
-FILE_PATTERN = "^{lab}/Subjects/{subject}/{date}/{number}/alf/{filename}$"
-FILE_REGEX = _pattern_to_regex(FILE_PATTERN)
-
-
-def _metadata(origin):
-    """
-    Metadata dictionary for Parquet files.
-
-    :param origin: path to full directory, or computer name / db name
-    """
-    return {
-        'date_created': datetime.datetime.now().isoformat(sep=' ', timespec='minutes'),
-        'origin': str(origin),
-    }
 
 
 # -------------------------------------------------------------------------------------------------
 # Parsing util functions
 # -------------------------------------------------------------------------------------------------
 
-def _ses_eid(rel_ses_path):
-    m = SESSION_REGEX.match(str(rel_ses_path))
-    if not m:
-        raise ValueError("The relative session path `%s` is invalid." % rel_ses_path)
-    out = {n: m.group(n) for n in ('lab', 'subject', 'date', 'number')}
-    return SESSION_PATTERN.format(**out)
+def _ses_str_id(session_path):
+    """Returns a str id from a session path in the form '(lab/)subject/date/number'"""
+    return Path(*filter(None, session_path_parts(session_path, assert_valid=True))).as_posix()
 
 
-def _parse_rel_ses_path(rel_ses_path):
+def _get_session_info(rel_ses_path):
     """Parse a relative session path."""
-    m = SESSION_REGEX.match(str(rel_ses_path))
-    if not m:
-        raise ValueError("The relative session path `%s` is invalid." % rel_ses_path)
-    out = {n: m.group(n) for n in ('lab', 'subject', 'date', 'number')}
-    out['id'] = SESSION_PATTERN.format(**out)
-    out['number'] = int(out['number'])
+    out = session_path_parts(rel_ses_path, as_dict=True, assert_valid=True)
+    out['id'] = _ses_str_id(rel_ses_path)
     out['date'] = pd.to_datetime(out['date']).date()
+    out['number'] = int(out['number'])
     out['task_protocol'] = ''
     out['project'] = ''
     return out
 
 
-def _get_file_rel_path(file_path):
-    """Get the lab/Subjects/subject/... part of a file path."""
-    file_path = str(file_path).replace('\\', '/')
-    # Find the relative part of the file path.
-    i = file_path.index('/Subjects')
-    if '/' not in file_path[:i]:
-        return file_path
-    i = file_path[:i].rindex('/') + 1
-    return file_path[i:]
-
-
-# -------------------------------------------------------------------------------------------------
-# Other util functions
-# -------------------------------------------------------------------------------------------------
-
-def _walk(root_dir):
-    """Iterate over all files found within a root directory."""
-    for p in sorted(Path(root_dir).rglob('*')):
-        yield p
-
-
-def _is_session_dir(path):
-    """Return whether a path is a session directory.
-
-    Example of a session dir: `/path/to/root/mainenlab/Subjects/ZM_1150/2019-05-07/001/`
-
-    """
-    return path.is_dir() and path.parent.parent.parent.name == 'Subjects'
-
-
-def _is_file_in_session_dir(path):
-    """Return whether a file path is within a session directory."""
-    if path.name.startswith('.'):
-        return False  # Ignore hidden files
-    return not path.is_dir() and '/Subjects/' in str(path.parent.parent.parent).replace('\\', '/')
-
-
-def _find_sessions(root_dir):
-    """Iterate over all session directories found in a root directory."""
-    for p in _walk(root_dir):
-        if _is_session_dir(p):
-            yield p
-
-
-def _find_session_files(full_ses_path):
-    """Iterate over all files in a session, and yield relative dataset paths."""
-    for p in _walk(full_ses_path):
-        if not (p.is_dir() or p.name.startswith('.')):  # Ignore folders and hidden files
-            yield p.relative_to(full_ses_path)
-
-
 def _get_dataset_info(full_ses_path, rel_dset_path, ses_eid=None, compute_hash=False):
-    rel_ses_path = _get_file_rel_path(full_ses_path)
+    rel_ses_path = get_alf_path(full_ses_path)
     full_dset_path = Path(full_ses_path, rel_dset_path).as_posix()
     file_size = Path(full_dset_path).stat().st_size
-    ses_eid = ses_eid or _ses_eid(rel_ses_path)
+    ses_eid = ses_eid or _ses_str_id(rel_ses_path)
     return {
         'id': Path(rel_ses_path, rel_dset_path).as_posix(),
         'eid': str(ses_eid),
         'session_path': str(rel_ses_path),
         'rel_path': Path(rel_dset_path).as_posix(),
-        # 'dataset_type': '.'.join(str(rel_dset_path).split('/')[-1].split('.')[:-1]),
         'file_size': file_size,
         'hash': md5(full_dset_path) if compute_hash else None,
         'exists': True
     }
-
-
-# -------------------------------------------------------------------------------------------------
-# Main functions
-# -------------------------------------------------------------------------------------------------
-
-def _make_sessions_df(root_dir):
-    rows = []
-    for full_path in _find_sessions(root_dir):
-        rel_path = _get_file_rel_path(full_path)
-        ses_info = _parse_rel_ses_path(rel_path)
-        assert set(ses_info.keys()) <= set(SESSIONS_COLUMNS)
-        rows.append(ses_info)
-    df = pd.DataFrame(rows, columns=SESSIONS_COLUMNS)
-    return df
-
-
-def _extend_datasets_df(df, root_dir, rel_ses_path, hash_files=False):
-    rows = []
-    for rel_dset_path in _find_session_files(root_dir / rel_ses_path):
-        full_ses_path = root_dir / rel_ses_path
-        file_info = _get_dataset_info(full_ses_path, rel_dset_path, compute_hash=hash_files)
-        assert set(file_info.keys()) <= set(DATASETS_COLUMNS)
-        rows.append(file_info)
-    if df is None:
-        df = pd.DataFrame(rows, columns=DATASETS_COLUMNS)
-    else:
-        df = df.append(rows, ignore_index=True, verify_integrity=True)
-    return df
-
-
-def _make_datasets_df(root_dir, hash_files=False):
-    df = None
-    # Go through all found sessions.
-    for full_path in _find_sessions(root_dir):
-        rel_ses_path = _get_file_rel_path(full_path)
-        # Append the datasets of each session.
-        df = _extend_datasets_df(df, root_dir, rel_ses_path, hash_files=hash_files)
-    return df
 
 
 def _rel_path_to_uuid(df, id_key='rel_path', base_id=None, drop_key=False):
@@ -245,7 +120,71 @@ def _ids_to_int(df_ses, df_dsets, drop_id=False):
         df_dsets.drop('eid', axis=1, inplace=True)
 
 
-def make_parquet_db(root_dir, out_dir=None, hash_ids=True, hash_files=False):
+# -------------------------------------------------------------------------------------------------
+# Main functions
+# -------------------------------------------------------------------------------------------------
+
+def _metadata(origin):
+    """
+    Metadata dictionary for Parquet files.
+
+    Parameters
+    ----------
+    origin : str, pathlib.Path
+        Path to full directory, or computer name / db name
+    """
+    return {
+        'date_created': datetime.datetime.now().isoformat(sep=' ', timespec='minutes'),
+        'origin': str(origin),
+    }
+
+
+def _make_sessions_df(root_dir) -> pd.DataFrame:
+    """
+    Given a root directory, recursively finds all sessions and returns a sessions DataFrame
+
+    Parameters
+    ----------
+    root_dir : str, pathlib.Path
+        The folder to look for sessions
+
+    Returns
+    -------
+    A pandas DataFrame of session info
+    """
+    rows = []
+    for full_path in iter_sessions(root_dir):
+        # Get the lab/Subjects/subject/date/number part of a file path
+        rel_path = get_alf_path(full_path)
+        # A dict of session info extracted from path
+        ses_info = _get_session_info(rel_path)
+        assert set(ses_info.keys()) <= set(SESSIONS_COLUMNS)
+        rows.append(ses_info)
+    df = pd.DataFrame(rows, columns=SESSIONS_COLUMNS)
+    return df
+
+
+def _make_datasets_df(root_dir, hash_files=False):
+    df = pd.DataFrame([], columns=DATASETS_COLUMNS)
+    # Go through sessions and append datasets
+    for session_path in iter_sessions(root_dir):
+        rows = []
+        for rel_dset_path in _iter_datasets(session_path):
+            file_info = _get_dataset_info(session_path, rel_dset_path, compute_hash=hash_files)
+            assert set(file_info.keys()) <= set(DATASETS_COLUMNS)
+            rows.append(file_info)
+        df = df.append(rows, ignore_index=True, verify_integrity=True)
+    return df
+
+
+def _iter_datasets(session_path):
+    """Iterate over all files in a session, and yield relative dataset paths."""
+    for p in sorted(Path(session_path).rglob('*.*')):
+        if not p.is_dir() and is_valid(p.name):
+            yield p.relative_to(session_path)
+
+
+def make_parquet_db(root_dir, out_dir=None, hash_ids=True, hash_files=False, lab=None):
     root_dir = Path(root_dir).resolve()
 
     # Make the dataframes.
@@ -255,6 +194,14 @@ def make_parquet_db(root_dir, out_dir=None, hash_ids=True, hash_files=False):
     # Add integer id columns
     if hash_ids and len(df_ses) > 0:
         _ids_to_int(df_ses, df_dsets, drop_id=True)
+
+    if lab:  # Fill in lab name field
+        assert not df_ses['lab'].any() or (df_ses['lab'] == 'lab').all(), 'lab name conflict'
+        df_ses['lab'] = lab
+
+    # Check any files were found
+    if df_ses.empty or df_dsets.empty:
+        warnings.warn(f'No {"sessions" if df_ses.empty else "datasets"} found', RuntimeWarning)
 
     # Output directory.
     out_dir = Path(out_dir or root_dir)