+++ one/tests/test_alyxclient.py
@@ -294,10 +294,21 @@ class TestDownloadHTTP(unittest.TestCase):
         self.test_data_uuid = '84116a5c-131e-40a2-92d4-62c4aaae5c52'  # OpenAlyx dataset
 
     def test_paginated_request(self):
+        """Check that paginated response object is returned upon making large queries"""
         rep = self.ac.rest('datasets', 'list')
         self.assertTrue(isinstance(rep, one.webclient._PaginatedResponse))
         self.assertTrue(len(rep) > 250)
-        self.assertTrue(len([d['hash'] for d in rep]) == len(rep))
+        # This fails when new records are added/removed from the remote db while iterating
+        # self.assertTrue(len([_ for _ in rep]) == len(rep))
+
+        # Test what happens when list changes between paginated requests
+        name = '0A' + str(random.randint(0, 10000))
+        # Add subject between calls
+        rep = self.ac.rest('subjects', 'list', limit=10, no_cache=True)
+        s = self.ac.rest('subjects', 'create', data={'nickname': name, 'lab': 'cortexlab'})
+        self.addCleanup(self.ac.rest, 'subjects', 'delete', id=s['nickname'])
+        with self.assertWarns(RuntimeWarning):
+            _ = rep[10]
 
     def test_update_url_params(self):
         url = f'{self.ac.base_url}/sessions?param1=foo&param2=&limit=5&param3=bar'