+++ one/tests/test_one.py
@@ -283,6 +283,12 @@ class TestONECache(unittest.TestCase):
         self.assertEqual(2, len(verifiable))
 
     def test_list_datasets(self):
+        # test filename
+        dsets = self.one.list_datasets(filename='_ibl_trials*')
+        self.assertEqual(len(dsets), 18)
+        dsets = self.one.list_datasets(filename='gnagnag')
+        self.assertEqual(len(dsets), 0)
+
         # Test no eid
         dsets = self.one.list_datasets(details=True)
         self.assertEqual(len(dsets), len(self.one._cache.datasets))
@@ -621,7 +627,7 @@ class TestOneAlyx(unittest.TestCase):
 
     @classmethod
     def setUpClass(cls) -> None:
-        cls.tempdir = util.set_up_env()
+        cls.tempdir = util.set_up_env(use_temp_cache=False)
         with mock.patch('one.params.iopar.getfile', new=partial(util.get_file, cls.tempdir.name)):
             # util.setup_test_params(token=True)
             cls.one = OneAlyx(
@@ -686,8 +692,7 @@ class TestOneAlyx(unittest.TestCase):
 
     def test_pid2eid(self):
         pid = 'b529f2d8-cdae-4d59-aba2-cbd1b5572e36'
-        with mock.patch('one.params.iopar.getfile', new=partial(util.get_file, self.tempdir.name)):
-            eid, collection = self.one.pid2eid(pid, query_type='remote')
+        eid, collection = self.one.pid2eid(pid, query_type='remote')
         self.assertEqual('fc737f3c-2a57-4165-9763-905413e7e341', eid)
         self.assertEqual('probe00', collection)
         with self.assertRaises(NotImplementedError):
@@ -779,7 +784,7 @@ class TestOneRemote(unittest.TestCase):
 
         # Test details=False, with eid
         dsets = self.one.list_datasets(eid, details=False, query_type='remote')
-        self.assertIsInstance(dsets, np.ndarray)
+        self.assertIsInstance(dsets, list)
         self.assertEqual(110, len(dsets))
 
         with self.assertWarns(Warning):